+++ EDMC-MarketScout/load.py
@@ -14,6 +14,8 @@ from typing import Any, Dict, Optional
 
 PLUGIN_NAME = "EDMC-MarketScout"
 PLUGIN_VERSION = "0.5.0"
+VERSION = PLUGIN_VERSION
+__version__ = PLUGIN_VERSION
 
 APP_MODULE: Any = None